requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
playwright>=1.49.0
pyyaml>=6.0.1
pydantic>=2.10.0
